src/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
opencv-python==4.10.0.84
numpy==1.26.4
pandas==2.2.2
torch==2.3.1
torchvision==0.18.1
paddlepaddle==2.6.1
paddleocr==2.8.1
ultralytics
python-multipart
pyjwt==2.6.0
bcrypt==4.0.1
mysql-connector-python==8.0.33
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0